+++ src/com/WildAmazing/marinating/Demigods/Deities/Titans/Oceanus.java
@@ -192,7 +192,7 @@ public class Oceanus implements Deity
 			LASTCHECK = timeSent;
 			if((DMiscUtil.getOnlinePlayer(getPlayerName()) != null) && DMiscUtil.getOnlinePlayer(getPlayerName()).getWorld().hasStorm())
 			{
-				int healamt = (int) Math.ceil(0.1 * Math.pow(DMiscUtil.getDevotion(getPlayerName(), getName()), 0.297));
+				double healamt = Math.ceil(0.1 * Math.pow(DMiscUtil.getDevotion(getPlayerName(), getName()), 0.297));
 				if(DMiscUtil.getHP(getPlayerName()) + healamt > DMiscUtil.getMaxHP(getPlayerName())) healamt = DMiscUtil.getMaxHP(getPlayerName()) - DMiscUtil.getHP(getPlayerName());
 				DMiscUtil.setHP(getPlayerName(), DMiscUtil.getHP(getPlayerName()) + healamt);
 			}
