+++ src/com/WildAmazing/marinating/Demigods/Listeners/DDamage.java
@@ -97,13 +97,13 @@ public class DDamage implements Listener
 
 	public static void syncHealth(Player p)
 	{
-		int current = DMiscUtil.getHP(p);
+		double current = DMiscUtil.getHP(p);
 		if(current < 1)
 		{ // if player should be dead
 			p.setHealth(0);
 			return;
 		}
-		double ratio = ((double) current) / DMiscUtil.getMaxHP(p);
+		double ratio = current / DMiscUtil.getMaxHP(p);
 		int disp = (int) Math.ceil(ratio * 20);
 		if(disp < 1) disp = 1;
 		p.setHealth(disp);
@@ -212,7 +212,7 @@ public class DDamage implements Listener
 		return (int) (Math.round(reduction));
 	}
 
-	public static int specialReduction(Player p, int amount)
+	public static double specialReduction(Player p, double amount)
 	{
 		if(DMiscUtil.getActiveEffectsList(p.getName()) == null) return amount;
 		if(DMiscUtil.getActiveEffectsList(p.getName()).contains("Invincible"))
@@ -303,7 +303,7 @@ public class DDamage implements Listener
 		else return null;
 	}
 
-	public static Boolean cancelSoulDamage(Player p, int damage)
+	public static Boolean cancelSoulDamage(Player p, double damage)
 	{
 		if(damage >= DMiscUtil.getHP(p))
 		{
